Replace old span processor with new
